python_backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
numpy==1.26.0
python-multipart==0.0.6
pydantic==2.3.0
google-genai==0.2.2
